add comment

+++ src/main/java/bupt/hbq/spring/objects/info/MapView.java
@@ -3,9 +3,9 @@ package bupt.hbq.spring.objects.info;
 public class MapView {
     private String warn;
     private String name;
-    private int value;
+    private float value;
     public MapView() {}
-    public MapView(String name,int value,String warn){
+    public MapView(String name,Float value,String warn){
         this.warn = warn;
         this.name = name;
         this.value = value;
@@ -22,10 +22,10 @@ public class MapView {
 	public void setName(String name) {
 		this.name = name;
 	}
-	public int getValue() {
+	public float getValue() {
 		return value;
 	}
-	public void setValue(int value) {
+	public void setValue(Float value) {
 		this.value = value;
 	}
